Update APIHandler interface and implementation***
