+++ apihandlers/jamfpro/jamfpro_api_handler_constants.go
@@ -12,3 +12,43 @@ const (
 	OAuthAuthenticationSupport         = true                            // OAuthAuthSuppport: A boolean to indicate if the API supports OAuth authentication.
 	OAuthWithCertAuthenticationSupport = false                           // OAuthWithCertAuthSuppport: A boolean to indicate if the API supports OAuth with client certificate authentication.
 )
+
+// GetDefaultBaseDomain returns the default base domain used for constructing API URLs.
+func (j *JamfAPIHandler) GetDefaultBaseDomain() string {
+	return DefaultBaseDomain
+}
+
+// GetOAuthTokenEndpoint returns the endpoint for obtaining an OAuth token. Used for constructing API URLs.
+func (j *JamfAPIHandler) GetOAuthTokenEndpoint() string {
+	return OAuthTokenEndpoint
+}
+
+// GetBearerTokenEndpoint returns the endpoint for obtaining a bearer token. Used for constructing API URLs.
+func (j *JamfAPIHandler) GetBearerTokenEndpoint() string {
+	return BearerTokenEndpoint
+}
+
+// GetTokenRefreshEndpoint returns the endpoint for refreshing an existing token. Used for constructing API URLs.
+func (j *JamfAPIHandler) GetTokenRefreshEndpoint() string {
+	return TokenRefreshEndpoint
+}
+
+// GetTokenInvalidateEndpoint returns the endpoint for invalidating an active token. Used for constructing API URLs.
+func (j *JamfAPIHandler) GetTokenInvalidateEndpoint() string {
+	return TokenInvalidateEndpoint
+}
+
+// GetAPIBearerTokenAuthenticationSupportStatus returns a boolean indicating if bearer token authentication is supported in the api handler.
+func (j *JamfAPIHandler) GetAPIBearerTokenAuthenticationSupportStatus() bool {
+	return BearerTokenAuthenticationSupport
+}
+
+// GetAPIOAuthAuthenticationSupportStatus returns a boolean indicating if OAuth authentication is supported in the api handler.
+func (j *JamfAPIHandler) GetAPIOAuthAuthenticationSupportStatus() bool {
+	return OAuthAuthenticationSupport
+}
+
+// GetAPIOAuthWithCertAuthenticationSupportStatus returns a boolean indicating if OAuth with client certificate authentication is supported in the api handler.
+func (j *JamfAPIHandler) GetAPIOAuthWithCertAuthenticationSupportStatus() bool {
+	return OAuthWithCertAuthenticationSupport
+}
